--- x-pack/solutions/security/plugins/security_solution/public/common/utils/translations.ts
@@ -1,34 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import { i18n } from '@kbn/i18n';
-
-export const ruleName = i18n.translate('xpack.securitySolution.selector.groups.ruleName.label', {
-  defaultMessage: 'Rule name',
-});
-export const userName = i18n.translate('xpack.securitySolution.selector.grouping.userName.label', {
-  defaultMessage: 'User name',
-});
-export const hostName = i18n.translate('xpack.securitySolution.selector.grouping.hostName.label', {
-  defaultMessage: 'Host name',
-});
-export const sourceIP = i18n.translate('xpack.securitySolution.selector.grouping.sourceIP.label', {
-  defaultMessage: 'Source IP',
-});
-export const sourceAddress = i18n.translate(
-  'xpack.securitySolution.selector.groups.sourceAddress.label',
-  {
-    defaultMessage: 'Source address',
-  }
-);
-
-export const destinationAddress = i18n.translate(
-  'xpack.securitySolution.selector.groups.destinationAddress.label',
-  {
-    defaultMessage: 'Destination address',
-  }
-);
