+++ x-pack/solutions/security/plugins/security_solution/public/detection_engine/rule_details_ui/pages/rule_details/translations.ts
@@ -73,3 +73,31 @@ export const DELETE_CONFIRMATION_BODY = i18n.translate(
     defaultMessage: 'This action will delete the rule. Click "Delete" to continue.',
   }
 );
+
+export const SOURCE_ADDRESS = i18n.translate(
+  'xpack.securitySolution.detectionEngine.ruleDetails.groups.sourceAddress',
+  {
+    defaultMessage: 'Source address',
+  }
+);
+
+export const USER_NAME = i18n.translate(
+  'xpack.securitySolution.detectionEngine.ruleDetails.groups.userName',
+  {
+    defaultMessage: 'User name',
+  }
+);
+
+export const HOST_NAME = i18n.translate(
+  'xpack.securitySolution.detectionEngine.ruleDetails.groups.hostName',
+  {
+    defaultMessage: 'Host name',
+  }
+);
+
+export const DESTINATION_ADDRESS = i18n.translate(
+  'xpack.securitySolution.detectionEngine.ruleDetails.groups.destinationAddress',
+  {
+    defaultMessage: 'Destination address',
+  }
+);
