+++ x-pack/solutions/security/plugins/security_solution/public/detections/components/alerts_table/grouping_settings/query_builder.test.ts
@@ -5,18 +5,20 @@
  * 2.0.
  */
 
-import { getAlertsGroupingQuery } from '.';
+import type { AlertsGroupingQueryParams } from '.';
+import { defaultGroupStatsAggregations, getAlertsGroupingQuery } from '.';
 import { getQuery } from './mock';
 
-let sampleData = {
+let sampleData: AlertsGroupingQueryParams = {
+  additionalFilters: [{ bool: { filter: [], must: [], must_not: [], should: [] } }],
   from: '2022-12-29T22:57:34.029Z',
-  to: '2023-01-28T22:57:29.029Z',
+  groupStatsAggregations: defaultGroupStatsAggregations,
   pageIndex: 0,
   pageSize: 25,
   runtimeMappings: {},
-  uniqueValue: 'aSuperUniqueValue',
   selectedGroup: 'kibana.alert.rule.name',
-  additionalFilters: [{ bool: { filter: [], must: [], must_not: [], should: [] } }],
+  uniqueValue: 'aSuperUniqueValue',
+  to: '2023-01-28T22:57:29.029Z',
 };
 
 describe('getAlertsGroupingQuery', () => {
