+++ x-pack/solutions/security/plugins/security_solution/public/detections/components/alerts_table/grouping_settings/query_builder.ts
@@ -6,15 +6,19 @@
  */
 
 import type { BoolQuery } from '@kbn/es-query';
-import type { NamedAggregation } from '@kbn/grouping';
-import { isNoneGroup, getGroupingQuery } from '@kbn/grouping';
+import { getGroupingQuery, isNoneGroup, type NamedAggregation } from '@kbn/grouping';
 import type { RunTimeMappings } from '../../../../sourcerer/store/model';
 
-interface AlertsGroupingQueryParams {
+export interface AlertsGroupingQueryParams {
   additionalFilters: Array<{
     bool: BoolQuery;
   }>;
   from: string;
+  /**
+   * Function that returns the group aggregations by field.
+   * This is then used to render values in the EuiAccordion `extraAction` section.
+   */
+  groupStatsAggregations: (field: string) => NamedAggregation[];
   pageIndex: number;
   pageSize: number;
   runtimeMappings: RunTimeMappings;
@@ -26,6 +30,7 @@ interface AlertsGroupingQueryParams {
 export const getAlertsGroupingQuery = ({
   additionalFilters,
   from,
+  groupStatsAggregations,
   pageIndex,
   pageSize,
   runtimeMappings,
@@ -37,9 +42,7 @@ export const getAlertsGroupingQuery = ({
     additionalFilters,
     from,
     groupByField: selectedGroup,
-    statsAggregations: !isNoneGroup([selectedGroup])
-      ? getAggregationsByGroupField(selectedGroup)
-      : [],
+    statsAggregations: !isNoneGroup([selectedGroup]) ? groupStatsAggregations(selectedGroup) : [],
     pageNumber: pageIndex * pageSize,
     runtimeMappings,
     uniqueValue,
@@ -47,177 +50,3 @@ export const getAlertsGroupingQuery = ({
     sort: [{ unitsCount: { order: 'desc' } }],
     to,
   });
-
-const getAggregationsByGroupField = (field: string): NamedAggregation[] => {
-  const aggMetrics: NamedAggregation[] = [
-    {
-      unitsCount: {
-        cardinality: {
-          field: 'kibana.alert.uuid',
-        },
-      },
-    },
-  ];
-  switch (field) {
-    case 'kibana.alert.rule.name':
-      aggMetrics.push(
-        ...[
-          {
-            description: {
-              terms: {
-                field: 'kibana.alert.rule.description',
-                size: 1,
-              },
-            },
-          },
-          {
-            countSeveritySubAggregation: {
-              cardinality: {
-                field: 'kibana.alert.severity',
-              },
-            },
-          },
-          {
-            severitiesSubAggregation: {
-              terms: {
-                field: 'kibana.alert.severity',
-              },
-            },
-          },
-          {
-            usersCountAggregation: {
-              cardinality: {
-                field: 'user.name',
-              },
-            },
-          },
-          {
-            hostsCountAggregation: {
-              cardinality: {
-                field: 'host.name',
-              },
-            },
-          },
-          {
-            ruleTags: {
-              terms: {
-                field: 'kibana.alert.rule.tags',
-              },
-            },
-          },
-        ]
-      );
-      break;
-    case 'host.name':
-      aggMetrics.push(
-        ...[
-          {
-            rulesCountAggregation: {
-              cardinality: {
-                field: 'kibana.alert.rule.rule_id',
-              },
-            },
-          },
-          {
-            countSeveritySubAggregation: {
-              cardinality: {
-                field: 'kibana.alert.severity',
-              },
-            },
-          },
-          {
-            severitiesSubAggregation: {
-              terms: {
-                field: 'kibana.alert.severity',
-              },
-            },
-          },
-          {
-            usersCountAggregation: {
-              cardinality: {
-                field: 'user.name',
-              },
-            },
-          },
-        ]
-      );
-      break;
-    case 'user.name':
-      aggMetrics.push(
-        ...[
-          {
-            rulesCountAggregation: {
-              cardinality: {
-                field: 'kibana.alert.rule.rule_id',
-              },
-            },
-          },
-          {
-            countSeveritySubAggregation: {
-              cardinality: {
-                field: 'kibana.alert.severity',
-              },
-            },
-          },
-          {
-            severitiesSubAggregation: {
-              terms: {
-                field: 'kibana.alert.severity',
-              },
-            },
-          },
-          {
-            hostsCountAggregation: {
-              cardinality: {
-                field: 'host.name',
-              },
-            },
-          },
-        ]
-      );
-      break;
-    case 'source.ip':
-      aggMetrics.push(
-        ...[
-          {
-            rulesCountAggregation: {
-              cardinality: {
-                field: 'kibana.alert.rule.rule_id',
-              },
-            },
-          },
-          {
-            countSeveritySubAggregation: {
-              cardinality: {
-                field: 'kibana.alert.severity',
-              },
-            },
-          },
-          {
-            severitiesSubAggregation: {
-              terms: {
-                field: 'kibana.alert.severity',
-              },
-            },
-          },
-          {
-            hostsCountAggregation: {
-              cardinality: {
-                field: 'host.name',
-              },
-            },
-          },
-        ]
-      );
-      break;
-    default:
-      aggMetrics.push({
-        rulesCountAggregation: {
-          cardinality: {
-            field: 'kibana.alert.rule.rule_id',
-          },
-        },
-      });
-  }
-  return aggMetrics;
-};
